added possessor, made the code use the asp.net core to go with changes i have made to the server to use asp.net core (those will be included later when i figure out how to strip out sensitive information) also added portal and function specific program codes (intrusion in this case) working on alert but i may need to change how inititive is handled so the fortress gets a turn, and likely make it that the fortress rolls and the player rolls and the fortress programs just go after the fortress just like the programs i summon go right after me.
the fortress will need to decide what it will do based on information the alerting program gives (that it spoted an enemy and where the enemy is located) and it needs to change its actions based on how long its been since a program spotted an intruder.

of course the ref will also be able to direct the programs from the website, and will honestly probably do better than any AI i write for the purpose.

i also added commands for patrol and deactivate. deactivating a program will put it back in memory where it can be summoned again (useful for programs that can't move) also the only way to re-rezz a de-rezzed program if you don't have the re-rezz utility.